+++ scim2_models/rfc7644/search_request.py
@@ -2,6 +2,8 @@ from enum import Enum
 from typing import List
 from typing import Optional
 
+from pydantic import field_validator
+
 from .message import Message
 
 
@@ -38,6 +40,28 @@ class SearchRequest(Message):
     start_index: Optional[int] = None
     """An integer indicating the 1-based index of the first query result."""
 
+    @field_validator("start_index")
+    @classmethod
+    def start_index_floor(cls, value: int) -> int:
+        """According to :rfc:`RFC7644 §3.4.2 <7644#section-3.4.2.4>,
+        start_index values less than 0 are interpreted as 0.
+
+        A value less than 1 SHALL be interpreted as 1.
+        """
+
+        return max(0, value)
+
     count: Optional[int] = None
     """An integer indicating the desired maximum number of query results per
     page."""
+
+    @field_validator("count")
+    @classmethod
+    def count_floor(cls, value: int) -> int:
+        """According to :rfc:`RFC7644 §3.4.2 <7644#section-3.4.2.4>, count
+        values less than 1 are interpreted as 1.
+
+        A value less than 1 SHALL be interpreted as 1.
+        """
+
+        return max(1, value)